[Improve][connector-clickhouse] move generate shard list to `ClickhouseSourceSplitEnumerator`.
