+++ seatunnel-connectors-v2/connector-clickhouse/src/main/java/org/apache/seatunnel/connectors/seatunnel/clickhouse/sink/file/ClickhouseTable.java
@@ -81,11 +81,7 @@ public class ClickhouseTable implements Serializable {
         if (distributedEngine != null) {
             return String.format("%s.%s", getLocalDatabase(), getLocalTableName());
         } else {
-            return getTableIdentifier();
+            return String.format("%s.%s", database, tableName);
         }
-    }
-
-    public String getTableIdentifier() {
-        return String.format("%s.%s", database, tableName);
     }
 }
